pyproject.toml
[project]
name = "floravision-ai"
version = "0.1.0"
description = "An intelligent plant health assistant powered by computer vision and agentic reasoning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "streamlit>=1.40.0",
    "ultralytics>=8.3.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "xhtml2pdf>=0.2.17",
    "langchain-groq>=1.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
